Refactor chatbot tests to use fixtures for API key and configuration, enhancing test clarity and maintainability
